requirements.txt
cx_oracle>=8.3.0
mysql_connector_repackaged>=0.3.1
psycopg2>=2.9.10
psycopg2_binary>=2.9.9
pyodbc>=5.2.0
python-dotenv>=1.1.0
setuptools>=68.0.0
mysql-connector-python>=9.3.0